refactor(playground): split playground-server.ts into a typed route table plus domain modules
Extracts route dispatch, the bundler, the file watcher, SSE broadcasting,
SSR-renderer loading, static-asset serving, and port scanning out of the
3,011-line playground-server.ts into their own domain modules:

- route-table.ts: RouteDefinition/RouteContext types + matchRoute(), and
  playground-server.ts's handleRequest becomes a two-line dispatcher over an
  18-row ROUTES table (replacing the ~386-line if-chain in the old
  handleRequest).
- traversal-guard.ts: resolveSafePath(), replacing three duplicated inline
  path-traversal guards with one shared, unit-tested function.
- http-responses.ts: notFound().
- build-artifacts-shared.ts, page-bundle.ts, fixture-bundle.ts,
  shell-bundle.ts, scenario-bundle.ts: the four bundle families' compile +
  lazy-build logic, plus the cross-family artifact cache/dedup state and
  findArtifactForFamily().
- file-watcher.ts: the fs.watch pipeline, the rebuild-generation state
  machine, and invalidateCachesForChange().
- sse-broadcast.ts: the SSE client set and triggerReload()/handleEventsRoute().
- ssr-renderer.ts: shell and page server-renderer loading, with the
  last-good-fallback and warmup-decision helpers.
- static-assets.ts: the three CSS route handlers, now sharing
  resolveSafePath() instead of a duplicated inline guard.
- port-scanner.ts: resolvePreferredPort()/createHttpServerOnAvailablePort().
- manifest-cache.ts: the manifest cache state shared between
  playground-server.ts's readers and file-watcher.ts's invalidator (avoids a
  circular import between the two).

playground-server.ts (1,348 lines, down from 3,011) remains the composition
root: manifest lookup, page rendering, and startup — the parts that
compose every domain module into one running server, per the issue's
explicit scope.

Also fixes component-sources.ts's hardcoded `componentName === 'review-editor'`
CSS-sidecar check to read from Editor's real package.json exports map via a
new resolveEditorStylesheetUrl(), and fixes the stale `/c/:name` route
comment and a stale "components-scope-exempt" watcher-invalidation comment
that both referenced the shell-chrome files cluster #1172 already deleted.

The /page/<id> route contract and ?snapshot=1 mode are unchanged — verified
via the full unit suite, `bun run vercel-build` (static-exports every
route), and `bun run validate:playground` (crawls all 201 component +
201 documentation routes plus the SSE reload path against a live server).

Closes #1175
